Remove requirement that a watcher is only registered to one store
A watcher can perfectly handle being registered to multiple stores at
once without it affecting the values it triggers a callback on.

Since state is only accessed from within the `(store) => {}` function,
which is always called when a watcher is subscribed to a store, it can
handle multiple stores at once.